Add validation for settings with the wrong value type Fix #90170
